% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDb.R
\name{checkDb}
\alias{checkDb}
\title{Checks Dimensions Database for errors}
\usage{
checkDb(db)
}
\arguments{
\item{db}{The database to be checked}
}
\value{
Dataframe with HDIM identifier, error type, verbatim entry, and suggested correction.
}
\description{
\code{checkDb} runs through the online database and returns a dataframe associated with specific errors
}
\details{
Developed specifically for the Dimensions in Biodiversity Evolab Database.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

