% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuration.R
\name{checkDuration}
\alias{checkDuration}
\title{Checks Dimensions Database for invalid BeatingDuration values}
\usage{
checkDuration(db)
}
\arguments{
\item{db}{the database to be checked}
}
\value{
Dataframe with HDIM identifier, beating duration error tag, verbatim entry, and suggested correction.
}
\description{
\code{checkDuration} runs through the online database and returns a dataframe corresponding to invalid beating duration entries in the database
}
\details{
The sum of all values in the BeatingDuration column for any given plot should equal 420.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

