% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEmpty.R
\name{checkEmpty}
\alias{checkEmpty}
\title{Checks Dimensions Database for invalid empty entries}
\usage{
checkEmpty(db)
}
\arguments{
\item{db}{the database to be checked}
}
\value{
Dataframe with HDIM identifier and empty entry error tag.
}
\description{
\code{checkEmpty} processes the online database and returns a dataframe corresponding to invalid empty entries in the database
}
\details{
Only relevant columns are fully checked for empty entries - factor columns contingent to certain collection methods are also considered. Misplaced entries are also checked.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

