% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMisspell.R
\name{checkMisspell}
\alias{checkMisspell}
\title{Checks Dimensions Database for misspelled empty entries}
\usage{
checkMisspell(db)
}
\arguments{
\item{db}{the database to be checked}
}
\value{
List of character vectors of HDIM numbers
}
\description{
\code{checkEmpty} runs through the online database and returns a list of HDIM numbers corresponding to misspelled entries in the database
}
\details{
Existing entries are compared against approved entries as listed in the synonym tables
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

