% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTime.R
\name{checkTime}
\alias{checkTime}
\title{Checks Dimensions Database for invalid time entries}
\usage{
checkTime(db)
}
\arguments{
\item{db}{The database for which time entries are to be checked}
}
\value{
Dataframe with HDIM identifier, invalid time error tag, verbatim entry, and suggested correction.
}
\description{
\code{checkTime} processes the online database and returns a dataframe corresponding to invalid time entries in the database
}
\details{
Time entries will be checked, based on column name, for correct date and hour:minute formatting. Entries in the Date column are checked to be within the range of dates between March 1, 2014 and January 1, 2016.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}
