% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colEventsURL.R
\docType{data}
\name{colEventsURL}
\alias{colEventsURL}
\title{URL for collection events}
\format{An object of class \code{character} of length 1.}
\usage{
colEventsURL
}
\description{
URL for collection events
}
\details{
URL for collection events
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{datasets}

