% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbChecker.R
\name{dbChecker}
\alias{dbChecker}
\title{Checks Dimensions Database for errors}
\usage{
dbChecker(url)
}
\arguments{
\item{url}{The url where the goodle drive database lives}
}
\value{
A multi-leveled list of HDIM numbers
}
\description{
\code{dbChecker} runs through the online database and returns a list of HDIM numbers associated with specific errors
}
\details{
Developed specifically for the Dimensions in Biodiversity Evolab Database.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

