% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdimURLS.R
\docType{data}
\name{colEventsURL}
\alias{colEventsURL}
\alias{synPlotURL}
\alias{synCollectURL}
\alias{synMethodURL}
\alias{synPlantURL}
\alias{synPitURL}
\alias{synWhereURL}
\title{URLs for collection events and synonym tables}
\format{An object of class \code{character} of length 1.}
\usage{
colEventsURL

synPlotURL

synCollectURL

synMethodURL

synPlantURL

synPitURL

synWhereURL
}
\description{
URLs for collection events and synonym tables
}
\details{
URLs for collection events and synonym tables
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{datasets}
