% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabels.R
\name{makeLabels}
\alias{makeLabels}
\title{Make collection event or specimen labels}
\usage{
makeLabels(hdim, dir = NULL, sheetName, repID = 1)
}
\arguments{
\item{hdim}{the unique identifier(s) to be printed}

\item{dir}{the directory in which to save the labels}

\item{sheetName}{the file name to give the sheet of labels}

\item{repID}{the number of labels for each unique identifier (can be a single value or 
a vector of length equal to length of \code{hdim})}
}
\value{
A data.frame of the google sheet
}
\description{
\code{makeLabels} reads the online database and prints labels requested by 
unique identifier (HDIM or EMEC number)
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
