% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGoogle.R
\name{readGoogle}
\alias{readGoogle}
\title{Read a google sheet}
\usage{
readGoogle(u)
}
\arguments{
\item{u}{the URL of the google sheet}
}
\value{
A data.frame of the google sheet
}
\description{
\code{readGoogle} reads a google sheet into a data.frame
}
\details{
The google sheet must be published as a csv.  The function forces strings to be read as characters, not factors
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
