% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDb.R
\name{checkDb}
\alias{checkDb}
\title{Checks Dimensions Database for errors}
\usage{
checkDb(db, match = "index")
}
\arguments{
\item{db}{The database to be checked}

\item{match}{The autocorrection method to be used with misspelled entries}
}
\value{
Dataframe with HDIM identifier, error type, verbatim entry, and suggested correction.
}
\description{
\code{checkDb} processes the online database and returns a dataframe of errors and suggested corrections
}
\details{
Developed specifically for the Dimensions in Biodiversity Evolab Database.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

