% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEmpty.R
\name{checkEmpty}
\alias{checkEmpty}
\title{Checks Dimensions Database for invalid empty entries}
\usage{
checkEmpty(db)
}
\arguments{
\item{db}{the database to be checked}
}
\value{
Dataframe with HDIM identifier and empy entry error tag.
}
\description{
\code{checkEmpty} runs through the online database and returns a dataframe corresponding to invalid empty entries in the database
}
\details{
Only certain columns are checked for empty entries, factor columns contingent to certain collection methods are considered.
}
\author{
Edward Greg Huang <edwardgh@berkeley.edu>
}

